etl/requirements.txt
psycopg2-binary
requests
python-dotenv
pydantic

